+++ tests/src/test/groovy/io/github/joke/objects/tests/BeanFieldTest.groovy
@@ -1,0 +1,34 @@
+package io.github.joke.objects.tests
+
+import io.github.joke.objects.tests.basic.PersonImpl
+import spock.lang.Specification
+
+import static org.junit.platform.commons.util.ReflectionUtils.HierarchyTraversalMode.TOP_DOWN
+import static org.junit.platform.commons.util.ReflectionUtils.findFields
+import static org.junit.platform.commons.util.ReflectionUtils.isNotFinal
+import static org.junit.platform.commons.util.ReflectionUtils.isNotStatic
+import static org.junit.platform.commons.util.ReflectionUtils.isPrivate
+
+class BeanFieldTest extends Specification {
+
+    def 'field attributes'() {
+        setup:
+        def field = findFields(PersonImpl, {it.name == 'name'}, TOP_DOWN).first()
+
+        expect:
+        isPrivate field
+        isNotFinal field
+        isNotStatic field
+    }
+
+    def 'field with interface type'() {
+        setup:
+        def field = findFields(PersonImpl, {it.name == 'parents'}, TOP_DOWN).first()
+
+        expect:
+        isPrivate field
+        isNotFinal field
+        isNotStatic field
+    }
+
+}
